lab3/requirements.txt
Flask==2.0.3
Jinja2==3.0.3
flask-login-0.6.3

